requirements.txt
fastapi
uvicorn
yfinance
acryl-datahub
python-dotenv
pandas
httpx